orsem-django/requirements-new.txt
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.40.0
cryptography==42.0.8
defusedxml==0.8.0rc2
Deprecated==1.2.14
Django==2.2.12
django-cloudinary-storage==0.3.0
django-cors-headers==3.11.0
django-environ==0.4.5
django-extensions==3.1.1
django-oauth-toolkit==2.3.0
django-rest-framework-social-oauth2==1.2.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
drf-jwt==1.19.0
environ==1.0
Faker==26.0.0
google-api-core==2.19.1
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
jwcrypto==1.5.1
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==1.7.1
pyparsing==3.1.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
python3-openid==3.2.0
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rest-social-auth==6.0.1
rsa==4.9
sentry-sdk==2.6.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.5.0
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==2.2.2
wrapt==1.16.0
